Move enlarge and shrink animation methods to UiUtils
